Remove Front css no longer needed
