+++ demo/front.css
@@ -1,8 +1,6 @@
 .front {
-  height: calc(100vh - 44px);
   margin: 0 auto;
   max-width: 50em;
-  overflow: auto;
   padding: 3em 2em 0;
 }
