+++ src/components/customer-resource-show/index.js
@@ -1,0 +1,1 @@
+export { default } from './customer-resource-show';
