+++ src/components/package-show/index.js
@@ -1,0 +1,1 @@
+export { default } from './package-show';
