+++ src/components/package-show/package-show.css
@@ -1,0 +1,7 @@
+.detail-container {
+  padding: 0 1rem;
+}
+
+.detail-container-header {
+  margin: 2rem 0;
+}
