+++ src/components/title-show/index.js
@@ -1,0 +1,1 @@
+export { default } from './title-show';
