+++ src/components/title-show/title-show.css
@@ -1,0 +1,7 @@
+.detail-container {
+  padding: 0 1rem;
+}
+
+.detail-container-header {
+  margin: 2rem 0;
+}
